fix 凸状況を分離した
