+++ src/convex/declare/list.ts
@@ -1,31 +1,12 @@
 import * as Discord from 'discord.js'
 import Settings from 'const-settings'
 import * as declare from './status'
-import * as plan from '../plan/list'
 import * as current from '../../io/current'
 import * as damageList from '../../io/damageList'
 import * as schedule from '../../io/schedule'
 import * as status from '../../io/status'
 import * as util from '../../util'
 import {AtoE, Current, Damage, Member, Plan} from '../../util/type'
-
-/**
- * 凸予定一覧を更新する
- * @param alpha ボス番号
- * @param state 現在の状況
- * @param channel 凸宣言のチャンネル
- */
-export const SetPlan = async (alpha: AtoE, state?: Current, channel?: Discord.TextChannel) => {
-  state ??= await current.Fetch()
-  channel ??= util.GetTextChannel(Settings.DECLARE_CHANNEL_ID[alpha])
-
-  const msg = await channel.messages.fetch(Settings.DECLARE_MESSAGE_ID[alpha].PLAN)
-  const plans = await schedule.Fetch()
-  const text = await plan.CreatePlanText(alpha, state.stage, plans)
-
-  // 凸予定の前2行を取り除いて結合
-  await msg.edit('凸予定\n```ts\n' + text.split('\n').slice(2).join('\n'))
-}
 
 /**
  * メンバーの凸宣言から凸宣言一覧を作る
