+++ tsconfig.json
@@ -38,7 +38,7 @@
     "src/convex/lapAndBoss.ts",
     "src/convex/manage.ts",
     "src/convex/over.ts",
-    "src/convex/situation.ts",
+    "src/convex/timeLimit.ts",
     "src/convex/declare/command.ts",
     "src/convex/declare/index.ts",
     "src/convex/declare/interaction.ts",
@@ -47,7 +47,6 @@
     "src/convex/plan/delete.ts",
     "src/convex/plan/edit.ts",
     "src/convex/plan/index.ts",
-    "src/convex/plan/list.ts",
     "src/convex/plan/update.ts",
     "src/convex/report/cancel.ts",
     "src/convex/report/index.ts",
@@ -56,7 +55,10 @@
     "src/convex/role/attendance.ts",
     "src/convex/role/index.ts",
     "src/convex/role/kill.ts",
-    "src/convex/timeLimit.ts",
+    "src/convex/situation/boss.ts",
+    "src/convex/situation/index.ts",
+    "src/convex/situation/plan.ts",
+    "src/convex/situation/report.ts",
     "src/io/bossTable.ts",
     "src/io/current.ts",
     "src/io/damageList.ts",
